Small refactor, add dev-reqs
