+++ dev-requirements.txt
@@ -1,0 +1,19 @@
+hydra-core
+wandb
+tree-sitter
+pygraphviz
+graphviz
+numpy
+ipython
+torch
+transformers
+tokenizers
+box2d
+Flask
+
+isort
+black
+flake8
+mypy
+pytest
+pydocstyle
