+++ elm/__init__.py
@@ -1,1 +1,1 @@
-from elm.elm_main import ELM
+from elm.elm import ELM
